requirements.txt
numpy~=1.23.5
langchain~=0.0.219
numba~=0.56.4
pandas~=2.0.2
pytz~=2023.3
beautifulsoup4~=4.12.2
requests~=2.28.1
faiss-cpu
duckduckgo-search
alpaca_trade_api
pandas_market_calendars
tabulate